Factor out render module
